requirements.txt
python-pptx==1.0.2
pandas==2.2.3
PyMuPDF==1.25.5
PyYAML==6.0.2
openpyxl==3.1.5
scikit-learn==1.6.1
matplotlib==3.10.1
pyarrow==19.0.1